+++ pkg/cmd/model/entity.go
@@ -1,0 +1,6 @@
+package model
+
+type Entity struct {
+	Id   string `json:"Id"`
+	Name string `json:"Name"`
+}